+++ packages/core-react/src/components/player/players.tsx
@@ -8,7 +8,11 @@ import {
 
 import { PlayerObjectFit, PlayerProps } from './Player';
 
-export type VideoPlayerProps<TElement, TPoster> = {
+export type VideoPlayerProps<
+  TElement,
+  TPoster,
+  TPlaybackPolicyObject extends object,
+> = {
   src: (HlsSrc | VideoSrc | Base64Src)[] | null;
   objectFit: PlayerObjectFit;
   width?: string | number;
@@ -17,17 +21,26 @@ export type VideoPlayerProps<TElement, TPoster> = {
   title?: string;
   muted?: boolean;
   priority?: boolean;
-  poster?: PlayerProps<TElement, TPoster>['poster'];
+  poster?: PlayerProps<TElement, TPoster, TPlaybackPolicyObject>['poster'];
   jwt?: string;
-  isCurrentlyShown?: PlayerProps<TElement, TPoster>['_isCurrentlyShown'];
+  isCurrentlyShown?: PlayerProps<
+    TElement,
+    TPoster,
+    TPlaybackPolicyObject
+  >['_isCurrentlyShown'];
   options?: ControlsOptions;
   onStreamStatusChange?: (isLive: boolean) => void;
   onMetricsError?: (error: Error) => void;
   onAccessControlError?: (error: Error) => void;
+  onError?: (error: Error) => void;
 };
 
-export type AudioPlayerProps<TElement, TPoster> = Omit<
-  VideoPlayerProps<TElement, TPoster>,
+export type AudioPlayerProps<
+  TElement,
+  TPoster,
+  TPlaybackPolicyObject extends object,
+> = Omit<
+  VideoPlayerProps<TElement, TPoster, TPlaybackPolicyObject>,
   'src' | 'poster'
 > & {
   src: AudioSrc[] | null;