+++ packages/core/src/providers/studio/index.ts
@@ -3,6 +3,8 @@ import * as tus from 'tus-js-client';
 import {
   StudioAsset,
   StudioAssetPatchPayload,
+  StudioCreateAssetArgs,
+  StudioCreateAssetUrlArgs,
   StudioCreateStreamArgs,
   StudioPlaybackInfo,
   StudioStream,
@@ -32,6 +34,7 @@ import {
   Metrics,
   MirrorSizeArray,
   PlaybackInfo,
+  PlaybackPolicy,
   Stream,
   StreamSession,
   UpdateAssetArgs,
@@ -59,11 +62,15 @@ export class StudioLivepeerProvider extends BaseLivepeerProvider {
   }
 
   async createStream(args: CreateStreamArgs): Promise<Stream> {
+    const playbackPolicy = this._getPlaybackPolicyMapped(args.playbackPolicy);
     const studioStream = await this._create<
       StudioStream,
       StudioCreateStreamArgs
     >('/stream', {
-      json: args,
+      json: {
+        ...args,
+        ...(playbackPolicy ? { playbackPolicy } : {}),
+      },
       headers: this._defaultHeaders,
     });
     return this._mapToStream(studioStream);
@@ -71,6 +78,7 @@ export class StudioLivepeerProvider extends BaseLivepeerProvider {
 
   async updateStream(args: UpdateStreamArgs): Promise<Stream> {
     const streamId = typeof args === 'string' ? args : args.streamId;
+    const playbackPolicy = this._getPlaybackPolicyMapped(args.playbackPolicy);
 
     await this._update(`/stream/${streamId}`, {
       json: {
@@ -91,13 +99,8 @@ export class StudioLivepeerProvider extends BaseLivepeerProvider {
               },
             }
           : {}),
-        ...(typeof args?.playbackPolicy?.type !== 'undefined'
-          ? {
-              playbackPolicy: {
-                type: args.playbackPolicy.type,
-              },
-            }
-          : {}),
+
+        ...(playbackPolicy ? { playbackPolicy } : {}),
       },
       headers: this._defaultHeaders,
     });
@@ -151,14 +154,34 @@ export class StudioLivepeerProvider extends BaseLivepeerProvider {
     // upload all assets and do not throw for failed
     const pendingAssetIds = await Promise.allSettled(
       sources.map(async (source, index) => {
+        const playbackPolicy = this._getPlaybackPolicyMapped(
+          source.playbackPolicy,
+        );
+
         if ((source as CreateAssetSourceUrl).url) {
           const createdAsset = await this._create<
             { asset: StudioAsset },
-            CreateAssetSourceUrl
+            StudioCreateAssetUrlArgs
           >('/asset/upload/url', {
             json: {
               name: source.name,
               url: (source as CreateAssetSourceUrl).url,
+              storage: source?.storage?.ipfs
+                ? {
+                    ipfs: {
+                      spec: {
+                        nftMetadata: source?.storage?.metadata ?? {},
+                        ...(source?.storage?.metadataTemplate
+                          ? {
+                              nftMetadataTemplate:
+                                source.storage.metadataTemplate,
+                            }
+                          : {}),
+                      },
+                    },
+                  }
+                : undefined,
+              ...(playbackPolicy ? { playbackPolicy } : {}),
             },
             headers: this._defaultHeaders,
           });
@@ -167,10 +190,26 @@ export class StudioLivepeerProvider extends BaseLivepeerProvider {
         } else {
           const uploadReq = await this._create<
             { tusEndpoint: string; asset: { id: string } },
-            { name: string }
+            StudioCreateAssetArgs
           >('/asset/request-upload', {
             json: {
               name: source.name,
+              storage: source?.storage?.ipfs
+                ? {
+                    ipfs: {
+                      spec: {
+                        nftMetadata: source?.storage?.metadata ?? {},
+                        ...(source?.storage?.metadataTemplate
+                          ? {
+                              nftMetadataTemplate:
+                                source.storage.metadataTemplate,
+                            }
+                          : {}),
+                      },
+                    },
+                  }
+                : undefined,
+              ...(playbackPolicy ? { playbackPolicy } : {}),
             },
             headers: this._defaultHeaders,
           });
@@ -363,6 +402,7 @@ export class StudioLivepeerProvider extends BaseLivepeerProvider {
 
   async updateAsset(args: UpdateAssetArgs): Promise<Asset> {
     const { assetId, name, storage } = args;
+    const playbackPolicy = this._getPlaybackPolicyMapped(args.playbackPolicy);
     await this._update<Omit<StudioAssetPatchPayload, 'assetId'>>(
       `/asset/${assetId}`,
       {
@@ -373,11 +413,16 @@ export class StudioLivepeerProvider extends BaseLivepeerProvider {
                 ipfs: {
                   spec: {
                     nftMetadata: storage?.metadata ?? {},
-                    nftMetadataTemplate: storage?.metadataTemplate ?? 'player',
+                    ...(storage?.metadataTemplate
+                      ? {
+                          nftMetadataTemplate: storage.metadataTemplate,
+                        }
+                      : {}),
                   },
                 },
               }
             : undefined,
+          ...(playbackPolicy ? { playbackPolicy } : {}),
         },
         headers: this._defaultHeaders,
       },
@@ -464,6 +509,22 @@ export class StudioLivepeerProvider extends BaseLivepeerProvider {
       type: 'ViewsMetrics',
       metrics: studioMetrics,
     };
+  }
+
+  _getPlaybackPolicyMapped(
+    policy: PlaybackPolicy | undefined,
+  ): PlaybackPolicy | null {
+    return policy && typeof policy?.type !== 'undefined'
+      ? policy.type === 'webhook'
+        ? {
+            type: policy.type,
+            webhookId: policy.webhookId,
+            webhookContext: policy.webhookContext,
+          }
+        : {
+            type: policy.type,
+          }
+      : null;
   }
 }