+++ packages/react/src/components/media/players/utils.ts
@@ -1,5 +1,5 @@
 const STREAM_OFFLINE_ERROR_MESSAGE = 'stream open failed';
-const ACCESS_CONTROL_ERROR_MESSAGE =
+export const ACCESS_CONTROL_ERROR_MESSAGE =
   'shutting down since this session is not allowed to view this stream';
 
 export const isStreamOfflineError = (error: Error): boolean =>